// backup/src/types.tsx
export type TGenderScope = 'L' | 'P' | 'ALL';
export type TJurusanScope = 'KITAB' | 'TAHFIDZ' | 'ALL';

// Role sesuai request (huruf kecil)
export type TRole = 'super_admin' | 'rois' | 'bendahara' | 'kesantrian' | 'dewan' | 'alumni';

// --- INTERFACES ---

export interface IProfile {
    id: string;
    email: string;
    role: TRole; // super_admin, bendahara, dll
    full_name: string;
    is_active: boolean;
    foto_url?: string;
    no_hp?: string;
    akses_gender: TGenderScope;
    akses_jurusan: TJurusanScope;
}